refactored to put config in its own package and classes - test for DB works fine, Test for HttpApi fails
