+++ src/main/kotlin/blog/BlogProperties.kt
@@ -1,6 +1,7 @@
 package blog
 
 import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.context.annotation.Configuration
 
 // TODO Use "val" instead of "lateinit var" when spring-boot#8762 will be fixed
 @ConfigurationProperties("blog")
