+++ src/main/kotlin/blog/HtmlController.kt
@@ -1,5 +1,7 @@
 package blog
 
+import blog.kotlin.format
+import blog.service.db.ArticleRepository
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.ui.set
@@ -8,8 +10,8 @@ import org.springframework.web.bind.annotation.PathVariable
 
 @Controller
 class HtmlController(private val repository: ArticleRepository,
-					 private val markdownConverter: MarkdownConverter,
-					 private val properties: BlogProperties) {
+                     private val markdownConverter: MarkdownConverter,
+                     private val properties: BlogProperties) {
 
 	@GetMapping("/")
 	fun blog(model: Model): String {
