+++ src/main/kotlin/blog/HttpApi.kt
@@ -1,11 +1,13 @@
 package blog
 
+import blog.service.db.ArticleRepository
+import blog.service.db.UserRepository
 import org.springframework.web.bind.annotation.*
 
 @RestController
 @RequestMapping("/api/article")
 class ArticleController(private val repository: ArticleRepository,
-						private val markdownConverter: MarkdownConverter) {
+                        private val markdownConverter: MarkdownConverter) {
 
 	@GetMapping("/")
 	fun findAll() = repository.findAllByOrderByAddedAtDesc()
