--- src/main/kotlin/blog/Repositories.kt
@@ -1,9 +1,0 @@
-package blog
-
-import org.springframework.data.repository.CrudRepository
-
-interface ArticleRepository : CrudRepository<Article, Long> {
-	fun findAllByOrderByAddedAtDesc(): Iterable<Article>
-}
-
-interface UserRepository : CrudRepository<User, String>
