+++ src/main/kotlin/blog/kotlin/Extensions.kt
@@ -1,4 +1,4 @@
-package blog
+package blog.kotlin
 
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatterBuilder
