+++ src/test/kotlin/blog/HttpApiTests.kt
@@ -1,5 +1,7 @@
 package blog
 
+import blog.service.db.ArticleRepository
+import blog.service.db.UserRepository
 import com.nhaarman.mockito_kotlin.whenever
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
